fix: align advanced message formatting cards with app styling
