+++ app/src/main/java/me/rerere/rikkahub/ui/pages/assistant/detail/AssistantPromptSubPage.kt
@@ -280,151 +280,6 @@ fun AssistantPromptSubPage(
 
 
         // ═══════════════════════════════════════════════════════════════════
-        // MESSAGE TEMPLATE
-        // ═══════════════════════════════════════════════════════════════════
-        Column(
-            modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp)
-        ) {
-            Surface(
-                modifier = Modifier.fillMaxWidth(),
-                color = if (me.rerere.rikkahub.ui.theme.LocalDarkMode.current) 
-                    MaterialTheme.colorScheme.surfaceContainerLow 
-                else 
-                    MaterialTheme.colorScheme.surfaceContainerHigh,
-                shape = me.rerere.rikkahub.ui.theme.AppShapes.CardLarge
-            ) {
-                Column(
-                    modifier = Modifier.padding(16.dp),
-                    verticalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    Text(
-                        text = stringResource(R.string.assistant_page_message_template),
-                        style = MaterialTheme.typography.titleMedium,
-                        color = MaterialTheme.colorScheme.onSurface
-                    )
-                    Text(
-                        text = stringResource(R.string.assistant_page_message_template_desc),
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                    Text(buildAnnotatedString {
-                        append(stringResource(R.string.assistant_page_template_variables_label))
-                        append(" ")
-                        append(stringResource(R.string.assistant_page_template_variable_role))
-                        append(": ")
-                        withStyle(SpanStyle(color = MaterialTheme.colorScheme.primary)) {
-                            append("{{ role }}")
-                        }
-                        append(", ")
-                        append(stringResource(R.string.assistant_page_template_variable_message))
-                        append(": ")
-                        withStyle(SpanStyle(color = MaterialTheme.colorScheme.primary)) {
-                            append("{{ message }}")
-                        }
-                        append(", ")
-                        append(stringResource(R.string.assistant_page_template_variable_time))
-                        append(": ")
-                        withStyle(SpanStyle(color = MaterialTheme.colorScheme.primary)) {
-                            append("{{ time }}")
-                        }
-                        append(", ")
-                        append(stringResource(R.string.assistant_page_template_variable_date))
-                        append(": ")
-                        withStyle(SpanStyle(color = MaterialTheme.colorScheme.primary)) {
-                            append("{{ date }}")
-                        }
-                    }, style = MaterialTheme.typography.bodySmall)
-                    
-                    DebouncedTextField(
-                        value = assistant.messageTemplate,
-                        onValueChange = {
-                            onUpdate(
-                                assistant.copy(
-                                    messageTemplate = it
-                                )
-                            )
-                        },
-                        stateKey = assistant.id,
-                        modifier = Modifier.fillMaxWidth(),
-                        minLines = 5,
-                        maxLines = 15,
-                        textStyle = LocalTextStyle.current.copy(
-                            fontSize = 12.sp,
-                            fontFamily = FontFamily.Monospace,
-                            lineHeight = 16.sp
-                        )
-                    )
-                    
-                    // Preview section
-                    Column(
-                        modifier = Modifier
-                            .clip(RoundedCornerShape(8.dp))
-                            .background(MaterialTheme.colorScheme.background)
-                            .padding(8.dp)
-                            .fillMaxWidth()
-                    ) {
-                        Text(
-                            text = stringResource(R.string.assistant_page_template_preview),
-                            style = MaterialTheme.typography.titleSmall
-                        )
-                        val rawMessages = listOf(
-                            UIMessage.user("Hello"),
-                            UIMessage.assistant("Hello, how can I help you?"),
-                        )
-                        val preview by produceState<UiState<List<UIMessage>>>(
-                            UiState.Success(rawMessages),
-                            assistant
-                        ) {
-                            value = runCatching {
-                                UiState.Success(
-                                    templateTransformer.transform(
-                                        ctx = TransformerContext(
-                                            context = context,
-                                            model = Model(modelId = "gpt-4o", displayName = "GPT-4o"),
-                                            assistant = assistant
-                                        ),
-                                        messages = rawMessages
-                                    )
-                                )
-                            }.getOrElse {
-                                UiState.Error(it)
-                            }
-                        }
-                        preview.onError {
-                            Text(
-                                text = it.message ?: it.javaClass.name,
-                                color = MaterialTheme.colorScheme.error
-                            )
-                        }
-                        preview.onSuccess {
-                            it.fastForEach { message ->
-                                val roleLabel = when (message.role) {
-                                    MessageRole.USER -> "User"
-                                    MessageRole.ASSISTANT -> "Assistant"
-                                    else -> message.role.name
-                                }
-                                val textContent = message.parts
-                                    .filterIsInstance<UIMessagePart.Text>()
-                                    .joinToString("\n") { it.text }
-                                Column(
-                                    modifier = Modifier.padding(vertical = 4.dp)
-                                ) {
-                                    Text(
-                                        text = "$roleLabel:",
-                                        style = MaterialTheme.typography.labelSmall,
-                                        color = MaterialTheme.colorScheme.primary
-                                    )
-                                    MarkdownBlock(content = textContent)
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-
-        // ═══════════════════════════════════════════════════════════════════
         // PRESET MESSAGES
         // ═══════════════════════════════════════════════════════════════════
         Column(
@@ -549,278 +404,8 @@ fun AssistantPromptSubPage(
             }
         }
 
-        // ═══════════════════════════════════════════════════════════════════
-        // REGEX RULES
-        // ═══════════════════════════════════════════════════════════════════
-        Column(
-            modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp)
-        ) {
-            Surface(
-                modifier = Modifier.fillMaxWidth(),
-                color = if (me.rerere.rikkahub.ui.theme.LocalDarkMode.current) 
-                    MaterialTheme.colorScheme.surfaceContainerLow 
-                else 
-                    MaterialTheme.colorScheme.surfaceContainerHigh,
-                shape = me.rerere.rikkahub.ui.theme.AppShapes.CardLarge
-            ) {
-                Column(
-                    modifier = Modifier.padding(16.dp),
-                    verticalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    Text(
-                        text = stringResource(R.string.assistant_page_regex_title),
-                        style = MaterialTheme.typography.titleMedium,
-                        color = MaterialTheme.colorScheme.onSurface
-                    )
-                    Text(
-                        text = stringResource(R.string.assistant_page_regex_desc),
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                    
-                    assistant.regexes.fastForEachIndexed { index, regex ->
-                        AssistantRegexCard(
-                            regex = regex,
-                            onUpdate = onUpdate,
-                            assistant = assistant,
-                            index = index
-                        )
-                    }
-                    Button(
-                        onClick = {
-                            onUpdate(
-                                assistant.copy(
-                                    regexes = assistant.regexes + AssistantRegex(
-                                        id = Uuid.random()
-                                    )
-                                )
-                            )
-                        },
-                        modifier = Modifier.fillMaxWidth()
-                    ) {
-                        Icon(Icons.Rounded.Add, null)
-                    }
-                }
-            }
-        }
 
 
-    }
-}
-
-@Composable
-private fun AssistantRegexCard(
-    regex: AssistantRegex,
-    onUpdate: (Assistant) -> Unit,
-    assistant: Assistant,
-    index: Int
-) {
-    var expanded by remember {
-        mutableStateOf(false)
-    }
-    Surface(
-        modifier = Modifier.fillMaxWidth(),
-        shape = RoundedCornerShape(10.dp),
-        color = MaterialTheme.colorScheme.background
-    ) {
-        Column(
-            modifier = Modifier
-                .padding(12.dp)
-                .animateContentSize(),
-            verticalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = regex.name,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                    modifier = Modifier
-                        .weight(1f)
-                        .widthIn(max = 200.dp)
-                )
-                HapticSwitch(
-                    checked = regex.enabled,
-                    onCheckedChange = { enabled ->
-                        onUpdate(
-                            assistant.copy(
-                                regexes = assistant.regexes.mapIndexed { i, reg ->
-                                    if (i == index) {
-                                        reg.copy(enabled = enabled)
-                                    } else {
-                                        reg
-                                    }
-                                }
-                            )
-                        )
-                    },
-                    modifier = Modifier.padding(start = 8.dp)
-                )
-                IconButton(
-                    onClick = {
-                        expanded = !expanded
-                    }
-                ) {
-                    Icon(
-                        imageVector = if (expanded) Icons.Rounded.KeyboardArrowUp else Icons.Rounded.KeyboardArrowDown,
-                        contentDescription = null
-                    )
-                }
-            }
-
-            if (expanded) {
-
-                DebouncedTextField(
-                    value = regex.name,
-                    onValueChange = { name ->
-                        onUpdate(
-                            assistant.copy(
-                                regexes = assistant.regexes.mapIndexed { i, reg ->
-                                    if (i == index) {
-                                        reg.copy(name = name)
-                                    } else {
-                                        reg
-                                    }
-                                }
-                            )
-                        )
-                    },
-                    label = stringResource(R.string.assistant_page_regex_name),
-                    stateKey = "regex_name_${regex.id}",
-                    modifier = Modifier.fillMaxWidth()
-                )
-
-                DebouncedTextField(
-                    value = regex.findRegex,
-                    onValueChange = { findRegex ->
-                        onUpdate(
-                            assistant.copy(
-                                regexes = assistant.regexes.mapIndexed { i, reg ->
-                                    if (i == index) {
-                                        reg.copy(findRegex = findRegex.trim())
-                                    } else {
-                                        reg
-                                    }
-                                }
-                            )
-                        )
-                    },
-                    label = stringResource(R.string.assistant_page_regex_find_regex),
-                    stateKey = "regex_find_${regex.id}",
-                    modifier = Modifier.fillMaxWidth()
-                )
-
-                DebouncedTextField(
-                    value = regex.replaceString,
-                    onValueChange = { replaceString ->
-                        onUpdate(
-                            assistant.copy(
-                                regexes = assistant.regexes.mapIndexed { i, reg ->
-                                    if (i == index) {
-                                        reg.copy(replaceString = replaceString)
-                                    } else {
-                                        reg
-                                    }
-                                }
-                            )
-                        )
-                    },
-                    label = stringResource(R.string.assistant_page_regex_replace_string),
-                    stateKey = "regex_replace_${regex.id}",
-                    modifier = Modifier.fillMaxWidth()
-                )
-
-                Column {
-                    Text(
-                        text = stringResource(R.string.assistant_page_regex_affecting_scopes),
-                        style = MaterialTheme.typography.labelMedium
-                    )
-                    Row(
-                        horizontalArrangement = Arrangement.spacedBy(16.dp)
-                    ) {
-                        AssistantAffectScope.entries.forEach { scope ->
-                            Row(
-                                verticalAlignment = Alignment.CenterVertically,
-                                horizontalArrangement = Arrangement.spacedBy(4.dp)
-                            ) {
-                                Checkbox(
-                                    checked = scope in regex.affectingScope,
-                                    onCheckedChange = { checked ->
-                                        val newScopes = if (checked) {
-                                            regex.affectingScope + scope
-                                        } else {
-                                            regex.affectingScope - scope
-                                        }
-                                        onUpdate(
-                                            assistant.copy(
-                                                regexes = assistant.regexes.mapIndexed { i, reg ->
-                                                    if (i == index) {
-                                                        reg.copy(affectingScope = newScopes)
-                                                    } else {
-                                                        reg
-                                                    }
-                                                }
-                                            )
-                                        )
-                                    }
-                                )
-                                Text(
-                                    text = scope.name.lowercase().replaceFirstChar { it.uppercase() },
-                                    style = MaterialTheme.typography.labelMedium
-                                )
-                            }
-                        }
-                    }
-                }
-
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    Checkbox(
-                        checked = regex.visualOnly,
-                        onCheckedChange = { visualOnly ->
-                            onUpdate(
-                                assistant.copy(
-                                    regexes = assistant.regexes.mapIndexed { i, reg ->
-                                        if (i == index) {
-                                            reg.copy(visualOnly = visualOnly)
-                                        } else {
-                                            reg
-                                        }
-                                    }
-                                )
-                            )
-                        }
-                    )
-                    Text(
-                        text = stringResource(R.string.assistant_page_regex_visual_only),
-                        style = MaterialTheme.typography.labelMedium
-                    )
-                }
-
-                TextButton(
-                    onClick = {
-                        onUpdate(
-                            assistant.copy(
-                                regexes = assistant.regexes.filterIndexed { i, _ ->
-                                    i != index
-                                }
-                            )
-                        )
-                    }
-                ) {
-                    Row(
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.spacedBy(4.dp),
-                    ) {
-                        Icon(Icons.Rounded.Delete, null)
-                        Text(stringResource(R.string.delete))
-                    }
-                }
-            }
-        }
     }
 }
